+++ packages/trello-work-cli/src/checklist.test.ts
@@ -52,6 +52,7 @@ async function card(): Promise<TrelloCard> {
     idList: 'list-inbox',
     dateLastActivity: 'v1',
     shortUrl: 'https://trello.com/c/AbCd1234/work-unit',
+    members: [],
   };
 }